+++ Cpp07/ex02/Array.tpp
@@ -53,7 +53,7 @@ T&	Array<T>::operator[](unsigned int index) const
 }
 
 template <typename T>
-size_t	Array<T>::size(void) const
+unsigned int	Array<T>::size(void) const
 {
 	return (_size);
 }
